moved extra functions to helpers instead and added test coverage
