+++ frontend/app/src/helpers/helpers-extended.ts
@@ -2,6 +2,25 @@
 import LighningDecoder from 'light-bolt11-decoder';
 import { getHost } from '../config/host';
 
+export const filterCount = (filterValues: any) => {
+  let count = 0;
+  for (const [, value] of Object.entries(filterValues)) {
+    if (value) {
+      count += 1;
+    }
+  }
+  return count;
+};
+
+export const formatSat = (sat: number) => {
+  if (sat === 0 || !sat) {
+    return '0';
+  }
+  const satsWithComma = sat.toLocaleString();
+  const splittedSat = satsWithComma.split(',');
+  return splittedSat.join(' ');
+};
+
 export const formatPrice = (amount = 0) => amount;
 
 export const formatSatPrice = (amount = 0) => {
